Settings: reconceive as a scope-grouped two-pane
Replace the 8-tab overflow-scrolling bar with a scope-grouped two-pane
(Account / Workspace / Developer / Moderation), grounded in /design +
settings-UX research — a reconception, not a reskin.

- New shared row grammar: SettingRow (label + only-if-helpful description
  left, control right) + SettingsGroup (hairline-divided rows on the surface,
  not card-per-setting). Adds settings-nav (container-query @2xl reflow: sticky
  grouped rail -> horizontal scroll-strip, one set of settings-tab-* testids)
  and settings-section (per-section voice header).
- Split Workspace into General (name/create/danger zone) + Members (typeahead
  + roster); add an Appearance section hosting the theme control; delete
  workspace-section.
- Drive the active section from the typed ?tab= search (getRouteApi) instead of
  the window.location hack; preserve the /settings?tab=github&gh_install
  redirect contract and the /settings/github/app/new server path.
- Restyle GitHub / Webhooks / Agents / Domains / Reports onto the new grammar;
  all integration logic (App install, live sync polling, delivery logs, DNS,
  tokens) and every inner testid preserved.
- Rename settings-tab-workspace -> settings-tab-general; add -members /
  -appearance; update smoke + deep specs for the split.

Verified: token/testid/frontend linters, tsc, biome, 89 unit tests, settings
smoke+deep+pr-previews e2e, and the screens harness (16 shots, both themes +
desktop/mobile).
