+++ apps/web/e2e/deep/settings.deep.spec.ts
@@ -1,22 +1,23 @@
 import { expect, test } from "../fixtures"
 
-// The Settings surface (tabbed: Workspace / Webhooks / Agents) in depth. The
-// `owner` fixture is the workspace owner (admin controls are owner-only); the
+// The Settings surface (Workspace › General / Members, Webhooks, Agents) in depth.
+// The `owner` fixture is the workspace owner (admin controls are owner-only); the
 // `secondUser` fixture supplies a real teammate to add as a member.
-test("owner manages workspace, webhooks, and agents through the tabs", async ({
+test("owner manages workspace, webhooks, and agents through the sections", async ({
   owner,
   secondUser,
 }) => {
   // Member / webhook / agent removal confirm via the in-app ConfirmDialog.
   await owner.goto("/settings")
 
-  // --- Workspace tab (Profile is the default now; switch to Workspace) ---
-  await owner.getByTestId("settings-tab-workspace").click()
+  // --- Workspace › General (Profile is the default; General holds the name) ---
+  await owner.getByTestId("settings-tab-general").click()
   await owner.getByTestId("workspace-name").fill("Acme HQ")
   await owner.getByTestId("workspace-save").click()
   await expect(owner.getByTestId("workspace-name")).toHaveValue("Acme HQ")
 
-  // Add the teammate as a commenter → owner + teammate = 2 member rows.
+  // --- Workspace › Members: add the teammate as a Viewer → owner + teammate = 2 rows. ---
+  await owner.getByTestId("settings-tab-members").click()
   await owner.getByTestId("member-email").fill(secondUser.email)
   await owner.getByTestId("member-role").click()
   await owner.getByRole("option", { name: "Viewer", exact: true }).click()
