+++ apps/web/e2e/smoke/settings.smoke.spec.ts
@@ -3,8 +3,8 @@ import { expect, test } from "../fixtures"
 // Settings loads for the owner and the workspace name persists through a save.
 test("settings loads for the owner and the workspace name saves", async ({ owner }) => {
   await owner.goto("/settings")
-  // Profile is the default tab now; switch to Workspace for the name field.
-  await owner.getByTestId("settings-tab-workspace").click()
+  // Profile is the default section now; switch to Workspace › General for the name.
+  await owner.getByTestId("settings-tab-general").click()
 
   await owner.getByTestId("workspace-name").fill("Acme HQ")
   await owner.getByTestId("workspace-save").click()
