+++ apps/web/src/pages/settings/index.tsx
@@ -1,31 +1,38 @@
+import { getRouteApi } from "@tanstack/react-router"
 import { useCallback, useEffect, useState } from "react"
 import { api, type Report } from "@/api"
-import { PageHeader } from "@/components/shared/page-header"
 import { PageShell } from "@/components/shared/page-shell"
+import { Eyebrow } from "@/components/shared/section-eyebrow"
 import { Badge } from "@/components/ui/badge"
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { useAuth } from "@/ctx"
 import { AgentsSection } from "./agents-section"
+import { AppearanceSection } from "./appearance-section"
 import { CustomDomainsSection } from "./custom-domains-section"
+import { GeneralSection } from "./general-section"
 import { GithubSection } from "./github-section"
 import { IntegrationsSection } from "./integrations-section"
+import { MembersSection } from "./members-section"
 import { ProfileSection } from "./profile-section"
 import { ReportsSection } from "./reports-section"
+import { SettingsNav, type SettingsNavGroup } from "./settings-nav"
 import { WebhooksSection } from "./webhooks-section"
-import { WorkspaceSection } from "./workspace-section"
 
+// The route owns the typed `?tab=` search; getRouteApi avoids a circular import
+// back to routes/settings.tsx while still giving typed useSearch/useNavigate.
+const route = getRouteApi("/settings")
+
+// Settings, reconceived as a scope-grouped two-pane: a sticky category rail
+// (Account · Workspace · Developer · Moderation) beside a readable detail column,
+// reflowing to a horizontal strip on a narrow pane. The active section rides the
+// URL's `?tab=` (deep-linkable, back-button-friendly) rather than the old local
+// window.location hack — and the GitHub App install still lands on
+// `/settings?tab=github&gh_install=…` unchanged. AppShell gates auth, so `me` is
+// present whenever Settings renders.
 export function Settings() {
-  // AppShell (mounted once around the Outlet) gates auth, so `me` is present
-  // whenever Settings renders.
   const { me } = useAuth()
   const [reports, setReports] = useState<Report[] | null>(null)
-  // Open straight to a tab when the URL asks for one (?tab=github) — the GitHub
-  // App install flow redirects back here with ?tab=github&gh_install=… so the
-  // section can finish in the repo picker. Defaults to the profile tab otherwise.
-  const [tab, setTab] = useState(() => {
-    if (typeof window === "undefined") return "profile"
-    return new URLSearchParams(window.location.search).get("tab") || "profile"
-  })
+  const { tab } = route.useSearch()
+  const nav = route.useNavigate()
 
   const loadReports = useCallback(
     () =>
@@ -38,85 +45,91 @@ export function Settings() {
   useEffect(() => {
     if (me) loadReports()
   }, [me, loadReports])
-  // The Reports tab only exists while there are open reports. If the last one is
-  // cleared while it's the active tab, fall back to Profile so the content panel
-  // never strands blank with no tab selected.
-  useEffect(() => {
-    if ((reports?.length ?? 0) === 0 && tab === "reports") setTab("profile")
-  }, [reports, tab])
 
   if (!me) return null
 
-  // Reports are urgent + owner-only — surface a tab only when there are open ones.
+  // Reports are urgent + owner-only — surface the section only while open ones exist.
   const openReports = reports ?? []
   const hasReports = openReports.length > 0
 
+  const groups: SettingsNavGroup[] = [
+    {
+      label: "Account",
+      items: [
+        { id: "profile", label: "Profile", testId: "settings-tab-profile" },
+        { id: "appearance", label: "Appearance", testId: "settings-tab-appearance" },
+      ],
+    },
+    {
+      label: "Workspace",
+      items: [
+        { id: "general", label: "General", testId: "settings-tab-general" },
+        { id: "members", label: "Members", testId: "settings-tab-members" },
+        { id: "integrations", label: "Integrations", testId: "settings-tab-integrations" },
+      ],
+    },
+    {
+      label: "Developer",
+      items: [
+        { id: "github", label: "GitHub", testId: "settings-tab-github" },
+        { id: "webhooks", label: "Webhooks", testId: "settings-tab-webhooks" },
+        { id: "agents", label: "Agents", testId: "settings-tab-agents" },
+        { id: "domains", label: "Domains", testId: "settings-tab-domains" },
+      ],
+    },
+  ]
+  if (hasReports) {
+    groups.push({
+      label: "Moderation",
+      items: [
+        {
+          id: "reports",
+          label: "Reports",
+          testId: "settings-tab-reports",
+          badge: (
+            <Badge variant="destructive" shape="pill">
+              {openReports.length}
+            </Badge>
+          ),
+        },
+      ],
+    })
+  }
+
+  // Guard the `?tab=` against unknown/stale values (e.g. `reports` after the last
+  // report clears) — fall back to Profile so the pane never strands blank.
+  const ids = groups.flatMap((g) => g.items.map((i) => i.id))
+  const active = tab && ids.includes(tab) ? tab : "profile"
+  const select = (id: string) => nav({ search: (prev) => ({ ...prev, tab: id }) })
+
   return (
-    // The shell's SidebarInset is already the page's <main> — this is just the
-    // scrolling content column.
-    <PageShell className="flex flex-col gap-6">
-      <PageHeader title="Settings" subtitle="Your profile, workspace, and integrations." />
+    <PageShell width="wide">
+      <div className="@container">
+        <div className="flex flex-col gap-5 @2xl:grid @2xl:grid-cols-[12rem_minmax(0,1fr)] @2xl:gap-8">
+          {/* The rail: page identity as a quiet mono eyebrow (the loud voice title
+              is the active section, not "Settings"), then the scope-grouped nav.
+              Sticky beside a scrolling detail column on a wide pane. */}
+          <div className="@2xl:sticky @2xl:top-8 @2xl:self-start">
+            <Eyebrow as="div" className="mb-3 hidden px-2 @2xl:block">
+              Settings
+            </Eyebrow>
+            <SettingsNav groups={groups} value={active} onSelect={select} />
+          </div>
 
-      <Tabs value={tab} onValueChange={setTab}>
-        <TabsList variant="line" className="max-w-full overflow-x-auto">
-          <TabsTrigger data-testid="settings-tab-profile" value="profile">
-            Profile
-          </TabsTrigger>
-          <TabsTrigger data-testid="settings-tab-workspace" value="workspace">
-            Workspace
-          </TabsTrigger>
-          <TabsTrigger data-testid="settings-tab-integrations" value="integrations">
-            Integrations
-          </TabsTrigger>
-          <TabsTrigger data-testid="settings-tab-webhooks" value="webhooks">
-            Webhooks
-          </TabsTrigger>
-          <TabsTrigger data-testid="settings-tab-agents" value="agents">
-            Agents
-          </TabsTrigger>
-          <TabsTrigger data-testid="settings-tab-github" value="github">
-            GitHub
-          </TabsTrigger>
-          <TabsTrigger data-testid="settings-tab-domains" value="domains">
-            Domains
-          </TabsTrigger>
-          {hasReports && (
-            <TabsTrigger data-testid="settings-tab-reports" value="reports">
-              Reports
-              <Badge variant="destructive" shape="pill">
-                {openReports.length}
-              </Badge>
-            </TabsTrigger>
-          )}
-        </TabsList>
-
-        <TabsContent value="profile">
-          <ProfileSection />
-        </TabsContent>
-        <TabsContent value="workspace">
-          <WorkspaceSection meId={me.id} />
-        </TabsContent>
-        <TabsContent value="integrations">
-          <IntegrationsSection />
-        </TabsContent>
-        <TabsContent value="webhooks">
-          <WebhooksSection />
-        </TabsContent>
-        <TabsContent value="agents">
-          <AgentsSection />
-        </TabsContent>
-        <TabsContent value="github">
-          <GithubSection />
-        </TabsContent>
-        <TabsContent value="domains">
-          <CustomDomainsSection />
-        </TabsContent>
-        {hasReports && (
-          <TabsContent value="reports">
-            <ReportsSection reports={openReports} reload={loadReports} />
-          </TabsContent>
-        )}
-      </Tabs>
+          <div className="min-w-0">
+            {active === "profile" && <ProfileSection />}
+            {active === "appearance" && <AppearanceSection />}
+            {active === "general" && <GeneralSection />}
+            {active === "members" && <MembersSection meId={me.id} />}
+            {active === "integrations" && <IntegrationsSection />}
+            {active === "github" && <GithubSection />}
+            {active === "webhooks" && <WebhooksSection />}
+            {active === "agents" && <AgentsSection />}
+            {active === "domains" && <CustomDomainsSection />}
+            {active === "reports" && <ReportsSection reports={openReports} reload={loadReports} />}
+          </div>
+        </div>
+      </div>
     </PageShell>
   )
 }
