+++ apps/web/src/pages/settings/profile-section.tsx
@@ -4,15 +4,18 @@ import { ProfileFields } from "@/components/profile-fields"
 import { AvatarPicker } from "@/components/shared/avatar-picker"
 import { FormField } from "@/components/shared/form-field"
 import { SectionTitle } from "@/components/shared/section-title"
-import { Card } from "@/components/ui/card"
-import { Checkbox } from "@/components/ui/checkbox"
+import { SettingRow } from "@/components/shared/setting-row"
+import { SettingsGroup } from "@/components/shared/settings-group"
+import { Switch } from "@/components/ui/switch"
 import { UsernameForm } from "@/components/username-form"
 import { useAuth } from "@/ctx"
 import { getInitials } from "@/lib/initials"
+import { SettingsSection } from "./settings-section"
 
-// Your personal profile (vs the Workspace tab): photo, handle, role + "what you
-// do", and people-search discoverability. The same fields onboarding collects, all
-// editable here. First tab in Settings — "we'll do even more in here" over time.
+// Your personal profile (vs. the Workspace sections): photo, handle, role +
+// "what you do", and people-search discoverability. The same fields onboarding
+// collects, all editable here — laid out as identity → role → a discoverability
+// toggle, on the page surface (no card stack).
 export function ProfileSection() {
   const { me, setMe } = useAuth()
   const [uploading, setUploading] = useState(false)
@@ -33,8 +36,7 @@ export function ProfileSection() {
     }
   }
 
-  const toggleDiscoverable = async () => {
-    const next = !discoverable
+  const toggleDiscoverable = async (next: boolean) => {
     setDiscoverable(next) // optimistic
     try {
       await api.setDiscoverable(next)
@@ -45,66 +47,59 @@ export function ProfileSection() {
   }
 
   return (
-    <section className="flex flex-col gap-6">
-      <p className="text-sm text-muted-foreground">
-        How you show up to your team across Derive: on your public profile, in the @mention picker,
-        and the member directory. Your email always stays private.
-      </p>
-
-      {/* Photo + handle */}
-      <Card className="p-4">
-        <div className="flex flex-wrap items-start gap-4">
-          <AvatarPicker
-            image={me.image}
-            initials={me.name ? initials : null}
-            uploading={uploading}
-            onPick={pickPhoto}
-            ariaLabel="Change your profile photo"
-            testId="profile-avatar"
+    <SettingsSection
+      title="Profile"
+      description="How you show up across Derive — your public profile, the @mention picker, and the member directory. Your email always stays private."
+    >
+      {/* Identity: photo + handle. */}
+      <div className="flex flex-wrap items-start gap-4">
+        <AvatarPicker
+          image={me.image}
+          initials={me.name ? initials : null}
+          uploading={uploading}
+          onPick={pickPhoto}
+          ariaLabel="Change your profile photo"
+          testId="profile-avatar"
+        />
+        <FormField
+          label="Username"
+          className="min-w-60 flex-1"
+          hint={
+            <>
+              <span className="font-medium text-foreground">{me.email}</span> stays private.
+            </>
+          }
+        >
+          <UsernameForm
+            initial={me.username ?? ""}
+            submitLabel={me.username ? "Update username" : "Save username"}
+            onClaimed={(username) => setMe({ ...me, username })}
           />
-          <FormField
-            label="Username"
-            className="min-w-60 flex-1"
-            hint={
-              <>
-                <span className="font-medium text-foreground">{me.email}</span> stays private.
-              </>
-            }
-          >
-            <UsernameForm
-              initial={me.username ?? ""}
-              submitLabel={me.username ? "Update username" : "Save username"}
-              onClaimed={(username) => setMe({ ...me, username })}
-            />
-          </FormField>
-        </div>
-      </Card>
+        </FormField>
+      </div>
 
-      {/* Role + what you do */}
-      <Card className="p-4">
-        <SectionTitle>Role</SectionTitle>
+      {/* Role + what you do — ProfileFields self-labels "Your role" / "What you do",
+          so the group heading groups them without echoing "Role". */}
+      <div className="flex flex-col gap-3">
+        <SectionTitle>About you</SectionTitle>
         <ProfileFields />
-      </Card>
+      </div>
 
-      {/* Discoverability */}
-      <Card className="p-4">
-        <SectionTitle>Discoverability</SectionTitle>
-        <label className="flex items-start gap-2.5 text-sm text-foreground">
-          <Checkbox
+      {/* Discoverability — an instant toggle, no save. */}
+      <SettingsGroup title="Discoverability">
+        <SettingRow
+          htmlFor="account-discoverable"
+          label="List me in people search"
+          description={`On by default. Your @${me.username ?? "handle"}, name, role, and photo appear in people search; turn it off to hide yourself.`}
+        >
+          <Switch
+            id="account-discoverable"
             data-testid="account-discoverable"
             checked={discoverable}
             onCheckedChange={toggleDiscoverable}
-            className="mt-0.5"
           />
-          <span>
-            Let people find me by username in search.
-            <span className="mt-0.5 block text-sm text-muted-foreground">
-              On by default. Your @{me.username ?? "handle"}, name, role, and photo show up in
-              people search; uncheck to hide yourself. Your email always stays private.
-            </span>
-          </span>
-        </label>
-      </Card>
-    </section>
+        </SettingRow>
+      </SettingsGroup>
+    </SettingsSection>
   )
 }
